More work on the logger, got the parser working (could still be buggy), the beginings of the plugin interface working, and the simple (stdin) front end working.  There will be 2 main front ends for the logger.
The UDP/Multicast logger which recives data via udp and has a TCP interface for stats, and the simple (stdin) logger which reads log lines via stdin and outputs the stats afterwards.

The UDP logger is for realtime stats and aggregation.
The stdin logger is for post analyzing logs and for debugging.

Started porting the functionality (and most of the code) from nxhitlogd to a plugin for the udplogger.  The plugin still needs alot of work, but it is begining to take shape. It uses the 'funroll-strings.pl' script to generate some of the token matching stuff.  The best example of how it works is nxhitlogd_module/extra_header.[ch] and nxhitlogd_module/nxhitlogd_module.c which matches extra_header.  With some of the benchs I did previously, the wacky generated trees match about 3-4 times faster then if (strcmp())/else trees/loops.